wip: isolate `NeynarFrameCard` into its own component
now need to separate so it works with a single frame / frame url
